requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asttokens==2.4.1
async-timeout==5.0.1
attrs==25.1.0
av==14.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comm==0.1.4
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.0
decorator==5.1.1
diffusers==0.32.2
dill==0.3.8
distro==1.9.0
einops==0.8.0
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.17.0
flash_attn==2.7.4.post1
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.26.0
ipython==8.17.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.5.0
jupyter_core==5.4.0
kiwisolver==1.4.8
lightning-utilities==0.12.0
madgrad==1.3
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.5.8
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.61.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==23.2
pandas==2.2.3
parso==0.8.3
peft==0.14.0
pexpect==4.8.0
pillow==11.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
propcache==0.2.1
protobuf==6.30.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.16.1
pyparsing==3.2.1
python-dateutil==2.8.2
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==25.1.1
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.32.3
ruamel.zip2tar==0.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
spicy==0.16.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.6.0
torchmetrics==1.6.1
torchvision==0.21.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.12.0
transformers==4.48.2
transformers-stream-generator==0.0.5
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.8
Werkzeug==3.1.3
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
